+++ packages/web3-utils/CHANGELOG.md
@@ -226,4 +226,4 @@ Documentation:
 ### Fixed
 
 - `toWei` support numbers in scientific notation (#6908)
-- `toWei` and `fromWei` now supports integers as a unit. (#7053)  
+- `toWei` and `fromWei` trims according to ether unit successfuly (#7044)